auth & jwt & protect middleaware aws done

+++ backend/controllers/goalControllers.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get goal
 // @route Get api/goals
 // @access Private
 getGoal = asyncHandler(async (req, res) => {
-  const goal = await Goal.find();
+  const goal = await Goal.find({ user: req.user.id });
   res.status(200).json(goal);
 });
 
@@ -19,6 +20,7 @@ setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -37,6 +39,18 @@ editGoal = asyncHandler(async (req, res) => {
     throw new Error("Couldnt find  this goal");
   }
 
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const editgoal = await Goal.findByIdAndUpdate(id, req.body);
 
   res.status(200).json(editgoal);
@@ -53,6 +67,18 @@ deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(500);
     throw new Error("Couldnt find  this goal");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const deletegoal = await Goal.findByIdAndDelete(id);
